Merged PR 21084: Check token expiry and re-route to setup if needed
## What's being changed

Impersonation tokens created for chat API spaces are short-lived by default. When patching chat profile data, we need to check the existing token expiry and re-route to setup to refresh the token if needed.

We're using chunks of Magento's Cardinal Commerce module code to decode the stored token without a key, and check its expiry.

## Why it's being changed

When customers start chats and then register or log in, or sign up for a newsletter, we try to patch the ongoing chat with some basic customer data (name and email) to assist the chat host in Engagement Cloud. Currently this only works for the first two hours after chat setup, so has rather limited benefit.

## How to review / test this change

- Clear any existing `api_space_id` and `api_token` config keys for Chat
- Clear cache
- Enable chat, save chat settings
- Start a chat on the front end while logged out
- Log in as a customer
- Confirm that your chat identifier in EC changes from 'Anonymous Contact' to your Magento customer name
- Now wait 3 hours
- Note your current stored `api_token`
- Go to an incognito window and start a new chat
- Repeat the other steps, checking that:
-- your token is updated in config data
-- the api space id stays the same
-- your EC chat is updated as before

Related work items: #114330