Merged PR 25973: Open chat buttons in a new tab
## What's being changed

In this PR we are opening Chat settings buttons in a new tab.

## Why it's being changed

In some browsers it is now not possible to render iframes that drop cookies. Currently in Safari, our links to manage Chat configuration and teams don't work as expected, because dotdigital.com requires cookies to handle authentication. The links load the correct iframe but it is not possible to login (either directly or via OAuth). Other browser vendors are expected to follow suit with this, so we're changing these iframe links to open in new tabs.

## How to review / test this change

- Go to Stores > Configuration > Developer and connect via oauth
https://dev.azure.com/dotdigital/ec/_git/ec-core-magento2-extension/pullrequest/25971 (checkout to this branch also)
- If you are connected chat buttons should be opening a new logged in EC tab in chat sections
- If not chat buttons should be opening a new login EC page

Related work items: #129493