Merged PR 27870: Regenerate API space id when saving chat config
## What's being changed

In this PR we regenerate api space id each time we save chat config

## Why it's being changed

Currently we use the ScopeConfigInterface to get the api space value.
But if api_space_id is already set for default level and not defined for any of the websites then the getValue() function returns the default one (although we call this with websiteScope).

## How to review / test this change

- In a multi website instance enable EC at default level and use different EC accounts for the rest of websites
- Enable chat at default level
- Enable chat for all the rest of the websites
- Make sure in core config data that you have api_space_id and api_token defined for all the websites
- Visit each website and make sure that your chat window is connected to the correct EC account

Related work items: #136731