Merged PR 28373: Add urlBase to chat embed code, remove launchTimeout
## What's being changed

- A `urlBase` property is being added to the chat widget configuration
- `launchTimeout` is being removed

## Why it's being changed

`urlBase` is now a required property in the embed code, and our implementation was missing it. `launchTimeout` has been removed because Magento loads the widget asynchronously anyway, and it always takes a second or two to appear after the main page has been rendered (so we don't really need further delay).

## How to review / test this change

- Test the widget loads

Related work items: #139730