Merged PR 28508: Change API space id check to catch null and undefined
## What's being changed

Our chatData.js widget script.

## Why it's being changed

There is a possibility - and this might occur when Chat is switched on and off inside of a 6-hour period, for example - for `chatData` to be present in customer data (i.e. browser local storage) but to be `null`. We're changing the condition to check for `null` in addition to checking for `undefined`, so that in this event the data will be reloaded.

## How to review / test this change

- In your default scope, switch off Chat
- Check the core_config_data is cleared of `chat_api_credentials`
- Clear cache
- In Magento front end, delete the local storage entry `mage-cache-storage` via Application > Local Storage
- Now re-save chat config
- Clear cache
- Reload the front end

Related work items: #140078