Merged PR 29368: Remove invalidate function to skip js error
## What's being changed

We're removing the `invalidate` call prior to refreshing the section in the stored customerData, in our chatData.js script.

## Why it's being changed

Occasionally we got a js error on frontend once chat iframe tries to load
![image.png](https://dev.azure.com/dotdigital/b2ba4bc1-4a9a-48f7-a69a-5cf84134b7c4/_apis/git/repositories/87899739-2326-460d-87ba-8d659f851b31/pullRequests/29368/attachments/image.png)

This has as a result chat iframe not to be visible:
![image (2).png](https://dev.azure.com/dotdigital/b2ba4bc1-4a9a-48f7-a69a-5cf84134b7c4/_apis/git/repositories/87899739-2326-460d-87ba-8d659f851b31/pullRequests/29368/attachments/image%20%282%29.png)

The console error occurs when the JS tries to `remove' a section that doesn't exist yet. It's only reproducible on 2.4+, and there is an open issue on Magento to resolve this: https://github.com/magento/magento2/issues/31920.

From our testing we decided that calling `invalidate` as well as `reload` on the next line was not necessary because both ultimately do the same thing (see notes below).

## How to review / test this change

- upgrade compile and deploy static content
- open storefront in incognito
- first try to replicate the issue in develop and then make sure that the issue resolves in this branch
- make sure that the removal of `invalidate` doesn't break the logic (remember chatData must be hard reloaded every 6hr's so to make sure that we don't fetch old conf's that later changed)

## Notes

- https://magento.stackexchange.com/questions/208223/how-to-force-reloading-of-a-customer-data-section
- https://magento.stackexchange.com/questions/239710/magento-2-what-is-the-different-between-invalidate-and-reload-if-they-both-upda

Related work items: #143276