Merged PR 29414: Add checks to confirm we have an apiSpaceId
## What's being changed

chatData.js. Before calling `startChat(chatData)`, we're confirming that we have an `apiSpaceId`.

## Why it's being changed

If Chat is enabled, but for some reason the chat_api_credentials/credentials/api_space_id config key is not set, the chat widget launches and sends a request to /apispaces/null/widget/config/combined - which 401s and leaves a console error.

## How to review / test this change

- Enable chat for a given scope
- In core_config_data, delete chat_api_credentials/credentials/api_space_id for the scope
- Clear cache
- On the front end, hard refresh
- Confirm that no console error is logged
- Confirm that an explanatory logline is logged to the console

Related work items: #143575