Merged PR 42072: Update Chat Profile controller to use correct interfaces
## What's being changed

We have removed a redundant class `Dotdigitalgroup\Chat\Model\EmailFlagManager`, and we have updated the `Dotdigitalgroup\Chat\Controller\Profile\Index` class to follow Magento best practices.

## Why it's being changed

This resolves some critical issues reported by the Magento upgrade compatibility tool.

## How to review / test this change

1. Enable chat
2. Test chat messaging works as expected.
3. Run DI compile insure there are no missing dependencies within any of the Dotdigital Modules

## Notes
The Flag Manager class was used in the very early days of the Chat codebase, when Chat was provided in the Email module. It provided a way of saving temporary config data during the trial setup. At the time we had to include it in our module in order to support Magento version 2.2 and lower (it was added to Magento in 2.3). However we stopped using it, and in fact it should never have even made it to v1.0.0 of the separate Chat module.

Related work items: #193594, #193596