Merged PR 48338: Remove chat teams button from settings
## What's being changed

The UI of the Chat Settings page - we removed the 'Configure Chat Teams' link.

## Why it's being changed

This button was a bit misleading because it landed the merchant/user directly in the general user management area of Dotdigital. It pertained to Chat only in so far as it allowed the enabling of user permissions for Chat; linking to the Chat team customisation area wasn't possible before because this was dependent on a team being created in the first place.

In any case we're now replacing it with a link to our Help Centre that explains everything.

## How to review / test this change

- Clear cache
- View link
- Visit link

Related work items: #227329