Merged PR 67624: Update module to rely on Chat served via Tag
## What's being changed

This PR adapts the Chat integration to use the widget as served by the Dotdigital Tag. Our module code no longer outputs a widget - it now only provides the chat profile update functionality.

Accordingly, the ChatWidgetTest MFTF test is removed.

## Why it's being changed

With the Dotdigital Tag, it is now possible to switch on Chat via 'Connect website' in Dotdigital. However, in order to capture and send profile updates for active chats, we still need to preserve our own module.

## How to review / test this change

- Turn off Chat in Magento, erase all chat-related config from core_config_data
- Use the tag-develop branch in Email
- Turn on Chat in Connect > Connect website
- See the widget appear in Dotdigital
- Turn on Chat profile updates in Magento
- Confirm that only one widget is loaded, with one set of network requests
- Start a chat
- Sign in as a customer (or register), or subscribe (it works with email capture)
- See a chat profile update sent to Dotdigital (set a breakpoint in Dotdigitalgroup\Chat\Controller\Profile\Index::execute() to monitor)

Related work items: #277743